+++ dephell/controllers/_mutator.py
@@ -1,4 +1,5 @@
 # built-in
+from itertools import product
 from logging import getLogger
 from typing import Iterable, Iterator, Optional, Sequence, Set, Tuple
 
@@ -8,12 +9,34 @@ import attr
 # app
 from ..config import config
 from ..models import Dependency, Group, RootDependency
-from ..utils import lazy_product
 from ._dependency import DependencyMaker
 from ._graph import Graph
 
 
 logger = getLogger('dephell.controllers')
+
+
+def lazy_product(*all_groups):
+    slices = [[] for _ in range(len(all_groups))]
+    all_groups = [iter(groups) for groups in all_groups]
+
+    while True:
+        has_tail = False
+        tail = []
+        for container, groups in zip(slices, all_groups):
+            group = next(groups, None)
+            tail.append(group)
+            if group is not None:
+                container.append(group)
+                has_tail = True
+        if not has_tail:
+            return
+
+        for groups in product(*slices):
+            for group, el in zip(groups, tail):
+                if el is not None and group == el:
+                    yield groups
+                    break
 
 
 @attr.s()
