+++ dephell/utils.py
@@ -1,28 +1,3 @@
-# built-in
-from itertools import product
-
-
-def lazy_product(*all_groups):
-    slices = [[] for _ in range(len(all_groups))]
-    all_groups = [iter(groups) for groups in all_groups]
-
-    while True:
-        has_tail = False
-        tail = []
-        for container, groups in zip(slices, all_groups):
-            group = next(groups, None)
-            tail.append(group)
-            if group is not None:
-                container.append(group)
-                has_tail = True
-        if not has_tail:
-            return
-
-        for groups in product(*slices):
-            for group, el in zip(groups, tail):
-                if el is not None and group == el:
-                    yield groups
-                    break
 
 
 # https://github.com/bottlepy/bottle/commit/fa7733e075da0d790d809aa3d2f53071897e6f76
